fixed float handling

+++ src/schema/TableQuery.php
@@ -186,7 +186,7 @@ class TableQuery implements \IteratorAggregate, \ArrayAccess, \Countable
             case 'int':
                 return (int)preg_replace('([^+\-0-9]+)', '', $value);
             case 'float':
-                return (int)preg_replace('([^+\-0-9.]+)', '', str_replace(',', '.', $value));
+                return (float)preg_replace('([^+\-0-9.]+)', '', str_replace(',', '.', $value));
             default:
                 return $value;
         }
